Translate the routes automatically.
This change makes the call to "MyApp::Application.routes.translate_from_file"
unnecessary. It's still possible to call this function for backwards
compatibility.

It also adds a new config option, "translation_file", which can be used instead
of the call to "translate_from_file" to change the default translation file.