Do not insert locale params if already present in path
Add some logic in RouteTranslator::Translator.translate_path to detect if the locale param is present
in path. If not, locale is inserted at the beginning of the path as before.

This change let developper controls where he want the local param to appear and allow use of more
complex routing as `/:country/:locale/path/to/some/pages`.