Fix #51 by mimicking rails route helpers
Instead of defining methods in the `TestCase` classes, add them to the
`route_set.named_routes.helpers` list and the testcases will find a way
to delegate it to the controller by looking up that list.

This is the way it works in `TestCase` internals, so we should try to be as close as
possible to it.

Also, change the way we test that the url helpers work on testcases,
instead of calling respond_to, we call them and rescue anything that's
not a NameError