Add verify_host_path_consistency option
When using different hosts to identify different locales, it's
desirable to ensure that only coherent url's work on the application.

Mismatching between the locale assigned to a domain (ej: example.es) and
the language of the path (ej: '/cesta-de-la-compra') should be considered
404's.

This commit adds a configuration option to enable this behavior.