Fix missing dup on blocks
The dup of the `blocks` argument was removed on the last refactor of the
route_set, but it was necessary. On the other hand, `defaults` doesn't need a
dup since this is already done inside
`::ActionDispatch::Routing::Mapper::Mapping.build`.

The new test showcases the bug (404 response) with the previous code.