Check for `.empty?` instead of `.any?`
`.any?` is not the opposite of `.empty?`, because the first one will
search for truthy values inside the array and is O(n), while the latter
only checks for array length and is O(1).

```
> [nil].any?
 => false 
> [nil].empty?
 => false 
```

This change may break use cases where `config.available_locales` only
contains falsy values, but that should be considered a bug in the
application.

Ref: https://stackoverflow.com/a/41325569