Add option for standard nested separator
This commit adds a new configuration option called `i18n_use_slash_separator`
to the Route Translator gem. This option controls whether or not to use the `/`
separator for nested route names in the gem's translation files.

The approach used in Route Translator was to remove the slash and prefer
a nested key. However, this commit changes the default separator to a slash
(`/`) character to match Rails' nested route syntax.

The old approach of using a nested key instead of a separator would have
prevented the following scenario:

```rb
Rails.application.routes.draw do
  localized do
    resources :appointments, only: :index do
      post :reservations, on: :member
    end
    namespace :appointments do
      resources :reservations, only: :index
    end
  end
end
```

because of the necessary double use of `reservations` under `appointments`.

Here's an example of how the new separator works in a translation file:

```yml
routes:
  controllers:
    people:
      products:
        favourites: fans

routes:
  controllers:
    people/products:
      favourites: fans
```

By default, the `i18n_use_slash_separator` option is set to false to preserve
backward compatibility with existing translation files.

However, if this option is set to `true`, Route Translator will use the slash
separator instead.

In the next major release of Route Translator, the `i18n_use_slash_separator`
option will be removed and the behavior will be forced to use the slash
separator.

Therefore, we recommend that users update their translation files to use the
slash separator and set the `i18n_use_slash_separator` option to `true` to
ensure compatibility with future versions of the gem.